+++ braintree/BTPayment/BTPaymentViewController.m
@@ -8,6 +8,8 @@
 #define BT_APP_COLOR             [UIColor clearColor]
 #define BT_APP_TEXT_COLOR        [UIColor colorWithWhite:85/255.0f alpha:1]
 
+#define VTCARDVIEW_TAG 9
+
 #define CELL_BACKGROUND_VIEW_TAG        10
 #define CELL_BACKGROUND_VIEW_SHADOW_TAG 11
 #define CELL_BORDER_COLOR               [[UIColor colorWithWhite:207/255.0f alpha:1] CGColor]
@@ -26,7 +28,7 @@
 #define SUBMIT_BUTTON_DOWN_PRESS_GRADIENT_START_COLOR [UIColor colorWithWhite:221/255.0f alpha:1]
 #define SUBMIT_BUTTON_DOWN_PRESS_GRADIENT_END_COLOR   [UIColor colorWithWhite:234/255.0f alpha:1]
 
-#define SUBMIT_BUTTON_GRADIENT_FRAME CGRectMake(0, 0, submitButton.frame.size.width, submitButton.frame.size.height)
+#define SUBMIT_BUTTON_GRADIENT_FRAME CGRectMake(0, 0, _submitButton.frame.size.width, _submitButton.frame.size.height)
 
 @interface BTPaymentViewController ()
 
@@ -36,7 +38,6 @@
 @property (strong, nonatomic) VTClient *client;
 @property (strong, nonatomic) BTPaymentActivityOverlayView *paymentActivityOverlayView;
 @property (strong, nonatomic) UIView *cellBackgroundView;
-@property (strong, nonatomic) UIView *paymentFormFooterView;
 @property (strong, nonatomic) UIButton *submitButton;
 
 @property (strong, nonatomic) UIView *disabledButtonGradientView;
@@ -46,21 +47,6 @@
 @end
 
 @implementation BTPaymentViewController
-
-// public
-@synthesize delegate;
-
-// private
-@synthesize venmoTouchEnabled;
-@synthesize hasPaymentMethods;
-@synthesize client;
-@synthesize paymentFormView;
-@synthesize cardView;
-@synthesize checkboxCardView;
-@synthesize paymentActivityOverlayView;
-@synthesize cellBackgroundView;
-@synthesize paymentFormFooterView;
-@synthesize submitButton;
 
 + (id)paymentViewControllerWithVenmoTouchEnabled:(BOOL)hasVenmoTouchEnabled {
     BTPaymentViewController *paymentViewController =
@@ -78,7 +64,7 @@
     }
 
     self.title = @"Payment";
-    self.venmoTouchEnabled = hasVenmoTouchEnabled;
+    _venmoTouchEnabled = hasVenmoTouchEnabled;
     _requestsZipInManualCardEntry = YES;
     return self;
 }
@@ -95,86 +81,77 @@
     self.tableView.backgroundView = nil;
     if (!_cornerRadius) _cornerRadius = BT_DEFAULT_CORNER_RADIUS;
     if (!_viewBackgroundColor) _viewBackgroundColor = [UIColor colorWithWhite:238/255.0f alpha:1];
-    self.viewBackgroundColor = _viewBackgroundColor; // Changes the display.
+    [self setViewBackgroundColor:_viewBackgroundColor]; // Changes the display.
 
-    if (self.venmoTouchEnabled) {
+    if (_venmoTouchEnabled) {
         Class class = NSClassFromString(@"VTClient");
         if (class) {
-            self.client = [class sharedClient];
-            self.client.delegate = self;
+            _client = [class sharedClient];
+            _client.delegate = self;
 
-            if (client.paymentMethodOptionStatus == VTPaymentMethodOptionStatusYes) {
-                hasPaymentMethods = YES;
+            if (_client.paymentMethodOptionStatus == VTPaymentMethodOptionStatusYes) {
+                _hasPaymentMethods = YES;
             }
         }
     }
 
     // Create the payment form
-    self.paymentFormView = [BTPaymentFormView paymentFormView];
-    self.paymentFormView.delegate = self;
-    self.paymentFormView.requestsZip = _requestsZipInManualCardEntry;
-    self.paymentFormView.backgroundColor = [UIColor clearColor];
+    _paymentFormView = [BTPaymentFormView paymentFormView];
+    _paymentFormView.delegate = self;
+    _paymentFormView.requestsZip = _requestsZipInManualCardEntry;
+    _paymentFormView.backgroundColor = [UIColor clearColor];
 
-    // Section footer view to display the VTCheckboxView view and manual card's submit button
-    paymentFormFooterView = [[UIView alloc] initWithFrame:
-                           CGRectMake(0, 0, self.view.frame.size.width,
-                                      (self.venmoTouchEnabled && self.client ? 120 : 60))];
-    paymentFormFooterView.backgroundColor = [UIColor clearColor];
-
-    if (self.venmoTouchEnabled && self.client) {
+    // Create the checkbox view, if requested.
+    if (_venmoTouchEnabled && _client) {
         // Set up the VTCheckboxView view
-        checkboxCardView = [self.client checkboxView];
-        [checkboxCardView setOrigin:CGPointMake(9, 0)]; // 10-1 for left-side alignment
-        [checkboxCardView setWidth:300];
-        [checkboxCardView setBackgroundColor:[UIColor clearColor]];
-        [checkboxCardView setTextColor:[UIColor grayColor]];
-        [paymentFormFooterView addSubview:checkboxCardView];
+        _checkboxView = [_client checkboxView];
+        [_checkboxView setOrigin:CGPointMake(-1, 0)]; // -1 for left-side alignment
+        [_checkboxView setBackgroundColor:[UIColor clearColor]];
+        [_checkboxView setTextColor:[UIColor grayColor]];
 
         [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(hideVTCardViewSection) name:UIApplicationWillResignActiveNotification object:nil];
     }
 
-    submitButton = [UIButton buttonWithType:UIButtonTypeCustom];
-    submitButton.frame = CGRectMake(10, paymentFormFooterView.frame.size.height - 50, 300, 40);
-    submitButton.autoresizingMask = UIViewAutoresizingFlexibleWidth;
-    [submitButton addSubview:self.normalButtonGradientView];
-    [submitButton bringSubviewToFront:submitButton.titleLabel];
-    submitButton.layer.cornerRadius = _cornerRadius;
-    submitButton.layer.borderWidth  = 1;
-    submitButton.layer.borderColor  = [SUBMIT_BUTTON_BORDER_DISABLED_COLOR CGColor];
-    submitButton.clipsToBounds = YES;
-    submitButton.titleLabel.font = [UIFont boldSystemFontOfSize:18];
-    [submitButton setTitle:@"Submit New Card" forState:UIControlStateNormal];
-    [submitButton setTitleColor:SUBMIT_BUTTON_NORMAL_TITLE_COLOR
+    _submitButton = [UIButton buttonWithType:UIButtonTypeCustom];
+    _submitButton.frame = CGRectMake(0, 0, 0, 40);
+    _submitButton.autoresizingMask = UIViewAutoresizingFlexibleWidth;
+    [_submitButton addSubview:self.normalButtonGradientView];
+    [_submitButton bringSubviewToFront:_submitButton.titleLabel];
+    _submitButton.layer.cornerRadius = _cornerRadius;
+    _submitButton.layer.borderWidth  = 1;
+    _submitButton.layer.borderColor  = [SUBMIT_BUTTON_BORDER_DISABLED_COLOR CGColor];
+    _submitButton.clipsToBounds = YES;
+    _submitButton.titleLabel.font = [UIFont boldSystemFontOfSize:18];
+    [_submitButton setTitle:@"Submit New Card" forState:UIControlStateNormal];
+    [_submitButton setTitleColor:SUBMIT_BUTTON_NORMAL_TITLE_COLOR
                        forState:UIControlStateNormal];
-    [submitButton setTitleColor:SUBMIT_BUTTON_DISABLED_TITLE_COLOR
+    [_submitButton setTitleColor:SUBMIT_BUTTON_DISABLED_TITLE_COLOR
                        forState:UIControlStateDisabled];
-    [submitButton addTarget:self action:@selector(submitCardInfo:)
+    [_submitButton addTarget:self action:@selector(submitCardInfo:)
            forControlEvents:UIControlEventTouchUpInside];
 
-    [submitButton addTarget:self action:@selector(submitButtonTouchDown)
+    [_submitButton addTarget:self action:@selector(submitButtonTouchDown)
            forControlEvents:UIControlEventTouchDown];
-    [submitButton addTarget:self action:@selector(submitButtonTouchDragExit)
+    [_submitButton addTarget:self action:@selector(submitButtonTouchDragExit)
            forControlEvents:UIControlEventTouchDragExit];
-    [submitButton addTarget:self action:@selector(submitButtonTouchDragEnter)
+    [_submitButton addTarget:self action:@selector(submitButtonTouchDragEnter)
            forControlEvents:UIControlEventTouchDragEnter];
 
-    UIView *topShadow = [[UIView alloc] initWithFrame:CGRectMake(0, 1, submitButton.frame.size.width, 1)];
+    UIView *topShadow = [[UIView alloc] initWithFrame:CGRectMake(0, 1, _submitButton.frame.size.width, 1)];
     topShadow.backgroundColor = [UIColor colorWithWhite:1 alpha:.1];
     topShadow.autoresizingMask = UIViewAutoresizingFlexibleWidth;
-    [submitButton addSubview:topShadow];
-
-    [paymentFormFooterView addSubview:submitButton];
-    submitButton.enabled = NO;
+    [_submitButton addSubview:topShadow];
+    _submitButton.enabled = NO;
 }
 
 - (void)viewDidAppear:(BOOL)animated {
     // If we know the user will have no option of seeing a VTCardView, then give firstResponder
     // to the BTPaymentFormView.
-    if ((self.venmoTouchEnabled && self.client &&
-        self.client.paymentMethodOptionStatus == VTPaymentMethodOptionStatusNo)
-        || !self.venmoTouchEnabled
-        || !self.client) {
-        [paymentFormView.cardNumberTextField becomeFirstResponder];
+    if ((_venmoTouchEnabled && _client &&
+        _client.paymentMethodOptionStatus == VTPaymentMethodOptionStatusNo)
+        || !_venmoTouchEnabled
+        || !_client) {
+        [_paymentFormView.cardNumberTextField becomeFirstResponder];
     }
 }
 
@@ -187,10 +164,10 @@
 // Hide keyboard when the user scrolls
 - (void)scrollViewWillBeginDragging:(UIScrollView *)scrollView {
     UITextField *firstResponder;
-    if ([paymentFormView.cardNumberTextField isFirstResponder]) firstResponder = paymentFormView.cardNumberTextField;
-    else if ([paymentFormView.monthYearTextField isFirstResponder]) firstResponder = paymentFormView.monthYearTextField;
-    else if ([paymentFormView.cvvTextField isFirstResponder]) firstResponder = paymentFormView.cvvTextField;
-    else if ([paymentFormView.zipTextField isFirstResponder]) firstResponder = paymentFormView.zipTextField;
+    if ([_paymentFormView.cardNumberTextField isFirstResponder]) firstResponder = _paymentFormView.cardNumberTextField;
+    else if ([_paymentFormView.monthYearTextField isFirstResponder]) firstResponder = _paymentFormView.monthYearTextField;
+    else if ([_paymentFormView.cvvTextField isFirstResponder]) firstResponder = _paymentFormView.cvvTextField;
+    else if ([_paymentFormView.zipTextField isFirstResponder]) firstResponder = _paymentFormView.zipTextField;
     if (firstResponder) {
         [firstResponder resignFirstResponder];
     }
@@ -199,7 +176,6 @@
 #pragma mark - Submit button states
 
 - (void)submitButtonTouchUpInside {
-//    submitButton.layer.borderColor  = [SUBMIT_BUTTON_BORDER_DISABLED_COLOR CGColor];
     [self swapSubmitButtonGradientViewTo:self.normalButtonGradientView from:self.pressedButtonGradientView];
 }
 
@@ -217,18 +193,18 @@
 
 - (void)swapSubmitButtonGradientViewTo:(UIView *)to from:(UIView *)from {
     [from removeFromSuperview];
-    [submitButton addSubview:to];
-    [submitButton bringSubviewToFront:submitButton.titleLabel];
+    [_submitButton addSubview:to];
+    [_submitButton bringSubviewToFront:_submitButton.titleLabel];
 }
 
 #pragma mark - BTPaymentViewController private methods
 
 - (void)prepareForDismissal {
-    [paymentActivityOverlayView dismissAnimated:YES];
+    [_paymentActivityOverlayView dismissAnimated:YES];
 }
 
 - (void)showErrorWithTitle:(NSString *)title message:(NSString *)message {
-    [paymentActivityOverlayView dismissAnimated:NO];
+    [_paymentActivityOverlayView dismissAnimated:NO];
     UIAlertView *alertView = [[UIAlertView alloc] initWithTitle:title message:message delegate:nil cancelButtonTitle:nil otherButtonTitles:@"OK", nil];
     [alertView show];
 }
@@ -236,34 +212,34 @@
 #pragma mark - BTPaymentViewController private methods
 
 - (void)submitCardInfo:(UIButton *)button {
-    if ([self.delegate respondsToSelector:@selector(paymentViewController:didSubmitCardWithInfo:andCardInfoEncrypted:)]) {
-        if (!paymentActivityOverlayView) {
-            paymentActivityOverlayView = [BTPaymentActivityOverlayView sharedOverlayView];
+    if ([_delegate respondsToSelector:@selector(paymentViewController:didSubmitCardWithInfo:andCardInfoEncrypted:)]) {
+        if (!_paymentActivityOverlayView) {
+            _paymentActivityOverlayView = [BTPaymentActivityOverlayView sharedOverlayView];
         }
-        [paymentActivityOverlayView show];
+        [_paymentActivityOverlayView show];
 
         // Get card info dictionary from the payment form.
-        NSDictionary *cardInfo = [self.paymentFormView cardEntry];
+        NSDictionary *cardInfo = [_paymentFormView cardEntry];
         NSDictionary *cardInfoEncrypted;
-        if (client && venmoTouchEnabled) {
+        if (_client && _venmoTouchEnabled) {
             // If Venmo Touch, encrypt card info with Braintree's CSE key
-            cardInfoEncrypted = [client encryptedCardDataAndVenmoSDKSessionWithCardDictionary:cardInfo];
+            cardInfoEncrypted = [_client encryptedCardDataAndVenmoSDKSessionWithCardDictionary:cardInfo];
         }
 
-        [self.delegate paymentViewController:self didSubmitCardWithInfo:cardInfo andCardInfoEncrypted:cardInfoEncrypted];
+        [_delegate paymentViewController:self didSubmitCardWithInfo:cardInfo andCardInfoEncrypted:cardInfoEncrypted];
     }
     [self submitButtonTouchUpInside];
 }
 
 - (void)paymentMethodFound {
-    if (hasPaymentMethods) {
+    if (_hasPaymentMethods) {
         // This case may happen when the user closes the app when viewing the payment form.
         // Open re-opening, [client refresh] will trigger (if no modal is visible) and the
         // cardView would not need animate in again if it already exists.
-        cardView = nil;
+        _cardView = nil;
         [self.tableView reloadData];
     } else {
-        hasPaymentMethods = YES;
+        _hasPaymentMethods = YES;
 
         [self.tableView insertSections:[NSIndexSet indexSetWithIndex:0]
                       withRowAnimation:UITableViewRowAnimationAutomatic];
@@ -274,7 +250,7 @@
 
 - (void)hideVTCardViewSection {
     if ([self.tableView numberOfSections] == 2) {
-        hasPaymentMethods = NO;
+        _hasPaymentMethods = NO;
         [self.tableView deleteSections:[NSIndexSet indexSetWithIndex:0]
                       withRowAnimation:UITableViewRowAnimationAutomatic];
         [self performSelector:@selector(reloadTitle) withObject:nil afterDelay:.3];
@@ -284,48 +260,50 @@
 - (void)reloadTitle {
     [self.tableView reloadData];
 
-    // webview is causing a crash when it exists in a table footer view and you reload a particular section
-    // look like a bug with Apple.
-    // http://stackoverflow.com/questions/11626572/uiwebview-as-view-for-footer-in-section-for-tableview-strange-crash-crash-log
-    //    [self.tableView reloadSections:[NSIndexSet indexSetWithIndex:(hasPaymentMethods ? 1 : 0)]
-    //                  withRowAnimation:UITableViewRowAnimationNone];
+// webview is causing a crash when it exists in a table footer view and you reload a particular section
+// look like a bug with Apple.
+// http://stackoverflow.com/questions/11626572/uiwebview-as-view-for-footer-in-section-for-tableview-strange-crash-crash-log
+//    [self.tableView reloadSections:[NSIndexSet indexSetWithIndex:(_hasPaymentMethods ? 1 : 0)]
+//                  withRowAnimation:UITableViewRowAnimationNone];
 }
 
 #pragma mark - UITableViewDelegate
 
 - (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath {
-    // 81 = height for VTCardView, 40 = height of enter card manually cell
-    return (hasPaymentMethods && indexPath.section == 0 ? 74 : 40);
+    if (_hasPaymentMethods && indexPath.section == 0) {
+        // VTCardView
+        return 74;
+    } else if (indexPath.row == 0) {
+        // BTPaymentFormView
+        return 40;
+    } else if (indexPath.row == 1) {
+        // VTCheckbox (if available) and Submit button
+        CGFloat height = ((_venmoTouchEnabled && _client ? _checkboxView.frame.size.height : 0)
+                          + _submitButton.frame.size.height);
+        return height;
+    }
+
+    return 0;
 }
 
 - (CGFloat)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section {
     return 40;
 }
 
-- (CGFloat)tableView:(UITableView *)tableView heightForFooterInSection:(NSInteger)section {
-    if (hasPaymentMethods && section == 0) {
-        // VTCardView
-        return 0.5f;
-    } else {
-        // BTPaymentFormView
-        return (self.venmoTouchEnabled ? 120 : 50);
-    }
-}
-
-- (UIView *)tableView:(UITableView *)tableView viewForFooterInSection:(NSInteger)section {
-    // VTCardView has no footer view.
-    // Payment form view has submit button and perhaps VTCheckboxView
-    return (hasPaymentMethods && section == 0 ? nil : paymentFormFooterView);
-}
-
 #pragma mark - UITableViewDataSource
 
 - (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView {
-    return (hasPaymentMethods ? 2 : 1);
+    return (_hasPaymentMethods ? 2 : 1);
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    return 1;
+    if (_hasPaymentMethods && section == 0) {
+        // VTCardView
+        return 1;
+    } else {
+        // BTPaymentFormView & VTCheckboxView + submit button
+        return 2;
+    }
 }
 
 // Don't use "tableView:titleForHeaderInSection:" because titles don't auto-update when
@@ -342,10 +320,10 @@
     titleLabel.shadowOffset = CGSizeMake(0, 1);
     [view addSubview:titleLabel];
     
-    if (hasPaymentMethods && section == 0) {
+    if (_hasPaymentMethods && section == 0) {
         titleLabel.text = @"Use a Saved Card";
     } else {
-        titleLabel.text = (hasPaymentMethods ? @"Or, Add a New Card" : @"Add a New Card");
+        titleLabel.text = (_hasPaymentMethods ? @"Or, Add a New Card" : @"Add a New Card");
     }
     
     return view;
@@ -356,12 +334,15 @@
 
     static NSString *UseCardCell         = @"UseCardCell";
     static NSString *PaymentFormViewCell = @"PaymentFormViewCell";
+    static NSString *PaymentFormViewFooterCell = @"PaymentFormViewFooterCell";
 
     NSString *currentCellIdentifier;
-    if (hasPaymentMethods && indexPath.section == 0) {
+    if (_hasPaymentMethods && indexPath.section == 0) {
         currentCellIdentifier = UseCardCell;
-    } else {
+    } else if (indexPath.row == 0) {
         currentCellIdentifier = PaymentFormViewCell;
+    } else {
+        currentCellIdentifier = PaymentFormViewFooterCell;
     }
 
     UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:UseCardCell];
@@ -373,87 +354,97 @@
         if ([currentCellIdentifier isEqualToString:PaymentFormViewCell]) {
             [self setUpPaymentFormViewForCell:cell];
         }
-    }
 
-    if ([currentCellIdentifier isEqualToString:UseCardCell]) {
-        [self setUpCardViewForCell:cell];
+        if ([currentCellIdentifier isEqualToString:UseCardCell]) {
+            [self setUpCardViewForCell:cell];
+        }
     }
     
     return cell;
 }
 
 - (void)tableView:(UITableView *)tableView willDisplayCell:(UITableViewCell *)cell forRowAtIndexPath:(NSIndexPath *)indexPath {
-    if (hasPaymentMethods && indexPath.section == 0) {
-        // Venmo Touch row
+    if (_hasPaymentMethods && indexPath.section == 0) {
+        // Venmo Touch row || checkbox + submit button row
         cell.backgroundView = nil;
     }
-    else {
+    else if (indexPath.row == 0) {
         // Customize the cell background view
         if (cell.backgroundView.tag != CELL_BACKGROUND_VIEW_TAG) {
-            cellBackgroundView = [[UIView alloc] initWithFrame:cell.frame];
-            cellBackgroundView.backgroundColor = [UIColor whiteColor];
-            cellBackgroundView.tag = CELL_BACKGROUND_VIEW_TAG;
-            cellBackgroundView.layer.cornerRadius  = _cornerRadius;
-            cellBackgroundView.layer.borderColor   = CELL_BORDER_COLOR;
-            cellBackgroundView.layer.borderWidth   = 1;
-            cellBackgroundView.layer.shadowRadius  = 1;
-            cellBackgroundView.layer.shadowOpacity = 1;
-            cellBackgroundView.layer.shadowColor   = [[UIColor whiteColor] CGColor];
-            cellBackgroundView.layer.shadowOffset  = CGSizeMake(0, 1);
-            cellBackgroundView.autoresizingMask = UIViewAutoresizingFlexibleWidth;
+            _cellBackgroundView = [[UIView alloc] initWithFrame:cell.frame];
+            _cellBackgroundView.backgroundColor = [UIColor whiteColor];
+            _cellBackgroundView.tag = CELL_BACKGROUND_VIEW_TAG;
+            _cellBackgroundView.layer.cornerRadius  = _cornerRadius;
+            _cellBackgroundView.layer.borderColor   = CELL_BORDER_COLOR;
+            _cellBackgroundView.layer.borderWidth   = 1;
+            _cellBackgroundView.layer.shadowRadius  = 1;
+            _cellBackgroundView.layer.shadowOpacity = 1;
+            _cellBackgroundView.layer.shadowColor   = [[UIColor whiteColor] CGColor];
+            _cellBackgroundView.layer.shadowOffset  = CGSizeMake(0, 1);
+            _cellBackgroundView.autoresizingMask = UIViewAutoresizingFlexibleWidth;
 
-            UIView *topShadowView = [[UIView alloc] initWithFrame:CGRectMake(3, 1, cellBackgroundView.frame.size.width, 1)];
+            UIView *topShadowView = [[UIView alloc] initWithFrame:CGRectMake(3, 1, _cellBackgroundView.frame.size.width, 1)];
             topShadowView.autoresizingMask = UIViewAutoresizingFlexibleWidth;
             topShadowView.backgroundColor = [UIColor colorWithWhite:0 alpha:.1];
             topShadowView.tag = CELL_BACKGROUND_VIEW_SHADOW_TAG;
-            [cellBackgroundView addSubview:topShadowView];
+            [_cellBackgroundView addSubview:topShadowView];
             [self adjustCellBackgroundViewShadow];
 
             cell.backgroundView = nil;
-            cell.backgroundView = cellBackgroundView;
+            cell.backgroundView = _cellBackgroundView;
 
             [self setUpPaymentFormViewForCell:cell];
         }
+    } else {
+        cell.backgroundView = nil;
+        CGFloat contentViewWidth = cell.contentView.frame.size.width;
+
+        [_checkboxView setWidth:contentViewWidth];
+        [cell.contentView addSubview:_checkboxView];
+
+        _submitButton.frame = CGRectMake(0, _checkboxView.frame.size.height, contentViewWidth, 40);
+        [cell.contentView addSubview:_submitButton];
     }
 }
 
 - (void)adjustCellBackgroundViewShadow {
     // Set the background cell's top shadow width.
-    UIView *topShadowView = [cellBackgroundView viewWithTag:CELL_BACKGROUND_VIEW_SHADOW_TAG];
+    UIView *topShadowView = [_cellBackgroundView viewWithTag:CELL_BACKGROUND_VIEW_SHADOW_TAG];
     CGFloat topShadowBuffer = ceilf(_cornerRadius/2.0f + (_cornerRadius > 10 ? 1 : 0)); //(_cornerRadius/2 - 1)
-    CGRect topShadowFrame = CGRectMake(topShadowBuffer, 1, cellBackgroundView.frame.size.width - topShadowBuffer*2, 1);
+    CGRect topShadowFrame = CGRectMake(topShadowBuffer, 1, _cellBackgroundView.frame.size.width - topShadowBuffer*2, 1);
     topShadowView.frame = topShadowFrame;
 }
 
 - (void)setUpCardViewForCell:(UITableViewCell *)cell {
-    if (!cardView) {
-        cardView = [self.client cardView];
+    if (!_cardView) {
+        _cardView = [_client cardView];
+        _cardView.tag = VTCARDVIEW_TAG;
         // Set styling defaults if they were set before cardView was initialized
-        if (_cornerRadius)                self.cardView.cornerRadius       = _cornerRadius;
+        if (_cornerRadius)                _cardView.cornerRadius           = _cornerRadius;
         if (_vtCardViewBackgroundColor)   self.vtCardViewBackgroundColor   = _vtCardViewBackgroundColor;
         if (_vtCardViewTitleFont)         self.vtCardViewTitleFont         = _vtCardViewTitleFont;
         if (_vtCardViewInfoButtonFont)    self.vtCardViewInfoButtonFont    = _vtCardViewInfoButtonFont;
     }
     
-    if (cardView && cell) {
-        [cardView setOrigin:CGPointMake(0, 0)];
-        [cardView setBackgroundColor:[UIColor clearColor]];
-        [cardView setWidth:300];
-        [cell.contentView addSubview:cardView];
+    if (_cardView && cell && ![cell.contentView viewWithTag:VTCARDVIEW_TAG]) {
+        [_cardView setOrigin:CGPointMake(0, 0)];
+        [_cardView setBackgroundColor:[UIColor clearColor]];
+        [_cardView setWidth:cell.frame.size.width];
+        [cell.contentView addSubview:_cardView];
     }
 }
 
 - (void)setUpPaymentFormViewForCell:(UITableViewCell *)cell {
-    [paymentFormView removeFromSuperview];
-    [cell.contentView addSubview:paymentFormView];
+    [_paymentFormView removeFromSuperview];
+    [cell.contentView addSubview:_paymentFormView];
 }
 
 
 #pragma mark - BTPaymentFormViewDelegate
 
 - (void)paymentFormView:(BTPaymentFormView *)paymentFormView didModifyCardInformationWithValidity:(BOOL)isValid {
-    submitButton.enabled = isValid;
-    submitButton.layer.borderColor  = (isValid ? [SUBMIT_BUTTON_BORDER_ENABLED_COLOR CGColor] : [SUBMIT_BUTTON_BORDER_DISABLED_COLOR CGColor]);
+    _submitButton.enabled = isValid;
+    _submitButton.layer.borderColor  = (isValid ? [SUBMIT_BUTTON_BORDER_ENABLED_COLOR CGColor] : [SUBMIT_BUTTON_BORDER_DISABLED_COLOR CGColor]);
 }
 
 #pragma mark - VTClientDelegate
@@ -465,8 +456,8 @@
         // Force tableview to reloadData, which renders VTCardView
         NSLog(@"payment method on file");
         [self paymentMethodFound];
-    } else if (hasPaymentMethods && paymentMethodOptionStatus != VTPaymentMethodOptionStatusYes) {
-        hasPaymentMethods = NO;
+    } else if (_hasPaymentMethods && paymentMethodOptionStatus != VTPaymentMethodOptionStatusYes) {
+        _hasPaymentMethods = NO;
         [self.tableView reloadData];
     }
 }
@@ -477,14 +468,14 @@
 
 - (void)client:(VTClient *)client approvedPaymentMethodWithCode:(NSString *)paymentMethodCode {
     // Return it to the delegate
-    if ([self.delegate respondsToSelector:
+    if ([_delegate respondsToSelector:
          @selector(paymentViewController:didAuthorizeCardWithPaymentMethodCode:)]) {
-        if (!paymentActivityOverlayView) {
-            paymentActivityOverlayView = [BTPaymentActivityOverlayView sharedOverlayView];
-            [paymentActivityOverlayView show];
+        if (!_paymentActivityOverlayView) {
+            _paymentActivityOverlayView = [BTPaymentActivityOverlayView sharedOverlayView];
+            [_paymentActivityOverlayView show];
         }
 
-        [delegate paymentViewController:self didAuthorizeCardWithPaymentMethodCode:paymentMethodCode];
+        [_delegate paymentViewController:self didAuthorizeCardWithPaymentMethodCode:paymentMethodCode];
     }
 }
 
@@ -496,7 +487,7 @@
 
 - (void)setRequestsZipInManualCardEntry:(BOOL)requestsZipInManualCardEntry {
     _requestsZipInManualCardEntry =
-    paymentFormView.requestsZip   = requestsZipInManualCardEntry;
+    _paymentFormView.requestsZip   = requestsZipInManualCardEntry;
 }
 
 - (void)setCornerRadius:(CGFloat)cornerRadius {
@@ -508,9 +499,9 @@
     }
 
     _cornerRadius = cornerRadius;
-    cardView.cornerRadius =
-    cellBackgroundView.layer.cornerRadius  =
-    submitButton.layer.cornerRadius  =
+    _cardView.cornerRadius =
+    _cellBackgroundView.layer.cornerRadius  =
+    _submitButton.layer.cornerRadius  =
     _cornerRadius;
 
     [self adjustCellBackgroundViewShadow];
@@ -523,17 +514,17 @@
 
 - (void)setVtCardViewBackgroundColor:(UIColor *)vtCardViewBackgroundColor {
     _vtCardViewBackgroundColor =
-    self.cardView.useCardButtonBackgroundColor = vtCardViewBackgroundColor;
+    _cardView.useCardButtonBackgroundColor = vtCardViewBackgroundColor;
 }
 
 - (void)setVtCardViewTitleFont:(UIFont *)vtCardViewTitleFont {
     _vtCardViewTitleFont =
-    self.cardView.useCardButtonTitleFont = vtCardViewTitleFont;
+    _cardView.useCardButtonTitleFont = vtCardViewTitleFont;
 }
 
 - (void)setVtCardViewInfoButtonFont:(UIFont *)vtCardViewInfoButtonFont {
     _vtCardViewInfoButtonFont =
-    self.cardView.infoButtonFont = vtCardViewInfoButtonFont;
+    _cardView.infoButtonFont = vtCardViewInfoButtonFont;
 }
 
 #pragma mark - UIButton Gradients